server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
python-dotenv==1.0.1
Sastrawi==1.0.1
cvxopt==1.3.2
numpy==2.0.2